Removing the cached global.ini.php from the git repo.
The file global.ini.php is in the .gitignore, but during the merge the file
was added and wasn't removed. Removing it now.

Ignoring the config/global.ini.php allows chaning the test database from
and to mysql and pgsql. It also allows for contributors who use a password
for the root account on the database.