Multi-database codebase in sync with upstream at #fc8ae5f.
Code base of the multi-database version is now in sync with upstream
at #fc8ae5f. Core and Plugins tests are passing but for some text
differences in queries. More tests failing in this codebase compared
to upstream code, more with the postgresql database when compated with
mysql.

Bug fixes:
\Piwik\Common::dummyWhenNotNumeric - Modified the pattern in the
preg_replace to replace -, +, . and digits instead of just digits.

\Piwik\Archive::uncompress - This function was using @gzuncompress
without checking if the database supports blob type. Corresponding
"compress" function in \Piwik\ArchiveProcessor checks for the blob
type support before using gzcompress on the value.

Changes:
Modified \Piwik\DataAccess\LogAggregator::queryActionsByDimension to
take another parameter $extraGroupBy, which if not false, will contain
columns to be added to the group-by clause in addition to the ones
added the queryActionsByDimension function.
Postgresql throws an error and halts when a non-aggregating column is
not present in the group by clause, mysql throws a warning and moves
on.

plugins/Actions/Archiver.php
   #archiveDayEntryActions
   #archiveDayExitActions
   #archiveDayActionsTime
added the $extraSelects columns to the group-by clause.

Replaced the mysql specific IF(bool, then, else) with COALESCE which
works in both mysql and postgresql. The if condition was as IS NULL
check, so COALESCE fits in.