Added the "hour" function to pgsql schema.
Added a plpgsql function "hour" to the postgres schema. This is an
alias to the EXTRACT(HOUR FROM xyz) function. Until now database
specific Generic class was used to get the appropriate function for
the database. The "generic" class requires a database connection.
The getSegmentsMetadata function of plugins/VisitTime/VisitTime.php
was being called before the database was setup while running the
Integration/TwoVisitsWithCustomVariables_SegmentMatchNONETest.php
test; which had the calls to $Generic->hour. Since the database
was not setup, it was failing with "database not found" error.

To avoid that, created the alias function for the postgresql schema
and replace the calls to $Generic->hour('xyz') with "HOUR(xyz)".