Integration tests passing with pgsql for #56ac168.
Error code for unique key violation was different on pgsql than
what was expected. Had to modify that on ArchiveTableCreator.

Tests involving RankingQuery were failing on pgsql due to a typo
in the LogVisit, "recoginze" instead of the expected "recognize".

Generic had to be modified to return the value as is for
booleans.

Segment.php (needed fields) was modified to return proper
values for select fields containing aggregate functions.