Models from Db\Factory.
Models were being instantiated directly with 'new Model()'. This makes
it difficult to have database specific models. Databases in question
being mysql and postgresql. Added a function to core/Db/Factory.php -
getModel, which takes namespace and name of the class as the arguments
to generate the model based on the database in use.

Now all the queries are either in Model or DAO classes. There are a few
exceptions for classes where the queries were (felt to be) database
agnostic.

Added a new interface Piwik\Db\FactoryCreated. All the models that have
queries implement this interface. It is just a placeholder without any
functions. Purpose is to help identify classes whose objects are being
generated with the Factory::getModel function. This can be helpful when
the Model classes too have to come from the dependency injection
container.